pyproject.toml
[tool.poetry]
name = "xai-gpt-shap-lima"
version = "0.1.0"
description = ""
authors = ["Jan Sernec (JanHuntersi) <jan.sernec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llvmlite = "^0.43"
scikit-learn = "^1.6.0"
openai = "^1.57.2"
numpy = "<2.1"
xgboost = "^2.1.3"
matplotlib = "^3.9.3"
python-dotenv = "^1.0.1"
certifi = ">=2024.7.4"
charset-normalizer = ">=3.3.2"
idna = ">=3.7"
markdown-it-py = ">=3.0.0"
mdurl = ">=0.1.2"
pygments = ">=2.17.2"
requests = ">=2.32.2"
rich = ">=13.7.0"
urllib3 = ">=2.2.2"
prompt-toolkit = ">=3.0.43"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
